+++ R/mergeFitbitData.R
@@ -3,11 +3,36 @@ mergeFitbitData = function(filenames = NULL, desiredtz = "", configtz = NULL) {
     stop("Provide at least two filenames")
   }
   cnt = 1
+  
+  # change order filenames such that
+  # heart rate if present is at the end, because heart rate 
+  # has irregular epoch length and needs to be interpolated
+  filenames = filenames[c(grep(pattern = "sleep", x = basename(filenames), invert = FALSE),
+                          grep(pattern = "calories", x = basename(filenames), invert = FALSE),
+                          grep(pattern = "step", x = basename(filenames), invert = FALSE),
+                          grep(pattern = "heart", x = basename(filenames), invert = FALSE))]
+  
   while (cnt <= length(filenames)) {
     D = readFitbit(filename = filenames[cnt], desiredtz = desiredtz, configtz = configtz)
     if (cnt == 1) {
       data = D
     } else {
+      if (length(grep("heart", x = colnames(D))) > 0) {
+        # interpolate heart rate to 30 seconds to ease merging with
+        # other data types
+        t0 = round(as.numeric(D$dateTime[1]) / 30) * 30
+        t1 = round(as.numeric(D$dateTime[nrow(D)]) / 30) * 30
+        newTime = seq(t0, t1, by = 30)
+        newD = as.data.frame(resample(raw = as.matrix(D$heart_rate), 
+                                      rawTime = as.numeric(D$dateTime),
+                                      time = newTime,
+                                      stop = nrow(D)))
+        colnames(newD)[1] = "heart_rate"
+        if (length(newTime) > nrow(newD)) newTime = newTime[1:nrow(newD)]
+        newD$dateTime = as.POSIXct(newTime, tz = desiredtz)
+        rm(D)
+        D = newD[, c("dateTime", "heart_rate")]
+      }
       # double names is possible when recording is split across json files
       # in that case there may be multiple calories, steps and sleep files
       doubleNames = colnames(D)[colnames(D) %in% colnames(data)] 
@@ -43,8 +68,7 @@ mergeFitbitData = function(filenames = NULL, desiredtz = "", configtz = NULL) {
     cnt = cnt + 1
   }
   data = data[order(data$dateTime),]
-  
-  # fill gaps
+  # fill gaps with NA values
   timeRange = range(data$dateTime)
   epochSize = min(diff(as.numeric(data$dateTime[1:pmin(10, nrow(data))])))
   timeFrame = data.frame(dateTime = seq( timeRange[1], timeRange[2], by = epochSize))